+++ src/components/tasks/task_detail_connected_component.tsx
@@ -2,7 +2,7 @@ import { Dispatch } from 'redux';
 import { Store } from '../../stores';
 import { updateTaskServicesAsync, UpdateTaskServicesAsync } from '../../stores/services';
 import { TaskDetailProps, TaskDetailActions, TaskDetailComponent, TaskServiceUpdater } from './task_detail';
-import { selectTaskServices, createRelatedServicesQueryFromTask } from '../../selectors/services';
+import { createRelatedServicesQueryFromTask } from '../../selectors/services';
 import {
     Id as TaskId, AddToSavedListAction, addToSavedList, ToggleCompletedAction,
     toggleCompleted, RemoveFromSavedListAction, removeFromSavedList,
@@ -13,6 +13,7 @@ import { RouterProps } from '../../application/routing';
 import { Task } from '../../selectors/tasks/task';
 import { pickSavedTaskIds } from '../../selectors/tasks/pick_saved_task_ids';
 import { TaskServices } from '../../selectors/services/task_services';
+import { selectTaskServices } from '../../selectors/services/select_task_services';
 
 interface StateProps extends TaskDetailProps {
     readonly searchQuery: string;
